cleaned up more stuff and removed hogging of one of the most essential components of the game
The MenuTransitionsHelper is a very special class to the game so I removed the upgrade on it. Object upgrading should be using sparringly and only when you need to completely override the behavior of a component while ensuring object injection and maintaining compatibility with both base game components and other mods.